Store refactoring, cleanup
