+++ src/client/store/reducers/authReducer.js
@@ -1,5 +1,5 @@
 import { SET_CURRENT_USER } from '../actions/types';
-import isEmpty from '../validation/is-empty';
+import isEmpty from '@client/validation/is-empty';
 
 const initialState = {
     isAuthenticated: false,
